Separate batch_job_info into a separate file.
